pyproject.toml
[project]
name = "production-api"
version = "0.1.0"
description = "Production-ready AI API with security, testing, cost optimization, and monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "langsmith>=0.7.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "langchain-anthropic>=1.4.1",
    "slowapi>=0.1.9",
    "pydantic-settings>=2.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
